feat(client): Update api to complete work with ts;
