+++ packages/solidarity-client/server/hasura/getAllIndividualUsers.ts
@@ -1,9 +1,8 @@
 import axios from 'axios'
 
-const query = `query {
-  solidarity_users(where: {longitude: {_is_null: false}, latitude: {_is_null: false}}) {
+const query = `query ($organization_id: bigint, $condition: String) {
+  solidarity_users(where: {organization_id: {_eq: $organization_id}, condition: {_eq: $condition}}) {
     address
-    organization_id
     user_id
     atendimentos_concludos_calculado_
     atendimentos_em_andamento_calculado_
@@ -20,13 +19,13 @@ const query = `query {
     whatsapp
     data_de_inscricao_no_bonde
     condition
-    tipo_de_acolhimento
   }
-}`
+}
+`
 
-const getAllUsers = async (organization_id, condition) => {
+const getAllIndividualUsers = async (organization_id, condition) => {
   const { HASURA_API_URL, X_HASURA_ADMIN_SECRET } = process.env
-  const response = await axios.post(HASURA_API_URL, {
+  const response = await axios.post(HASURA_API_URL || '', {
     query,
     variables: {
       organization_id, condition
@@ -45,4 +44,4 @@ const getAllUsers = async (organization_id, condition) => {
   }
 }
 
-export default getAllUsers
+export default getAllIndividualUsers
