+++ packages/solidarity-client/server/hasura/getAllVolunteerTickets.ts
@@ -1,23 +1,22 @@
 import axios from 'axios'
 
-const query = `query ($organization_id: bigint, $status_acolhimento: String) {
-  solidarity_tickets(where: {status_acolhimento: {_eq: $status_acolhimento}, organization_id: {_eq: $organization_id}}) {
+const query = `query ($organization_id: bigint) {
+  solidarity_tickets(where: {organization_id: {_eq: $organization_id}}) {
     requester_id
-    status_acolhimento
+    status_inscricao
     ticket_id
-    subject
     created_at
     status
+    subject
   }
 }`
 
-const getAllTickets = async (organization_id, status_acolhimento) => {
+const getAllTickets = async (organization_id) => {
   const { HASURA_API_URL, X_HASURA_ADMIN_SECRET } = process.env
-  const response = await axios.post(HASURA_API_URL, {
+  const response = await axios.post(HASURA_API_URL || '', {
     query,
     variables: {
-      organization_id,
-      status_acolhimento
+      organization_id
     }
   }, {
     headers: {
