+++ packages/solidarity-client/server/index.ts
@@ -6,13 +6,13 @@ import path from 'path'
 import body_parser from 'body-parser'
 require('dotenv-safe').config()
 
-// import api from './api'
-// import forward from './forward'
+import api from './api'
+import forward from './forward'
 
 // Assert required enviroment variables for app
-// assert(process.env.ZENDESK_API_USER !== undefined, 'Required enviroment variable ZENDESK_API_USER')
-// assert(process.env.ZENDESK_API_TOKEN !== undefined, 'Required enviroment variable ZENDESK_API_TOKEN')
-// assert(process.env.PORT !== undefined, 'Required enviroment PORT')
+assert(process.env.ZENDESK_API_USER !== undefined, 'Required enviroment variable ZENDESK_API_USER')
+assert(process.env.ZENDESK_API_TOKEN !== undefined, 'Required enviroment variable ZENDESK_API_TOKEN')
+assert(process.env.PORT !== undefined, 'Required enviroment PORT')
 
 const asyncMiddleware = fn => (req, res, next) => {
   Promise.resolve(fn(req, res, next))
@@ -31,8 +31,8 @@ app.use((err, req, res, next) => {
   res.status(500).json({ message: 'an error occurred' })
 })
 
-// app.get('/api', asyncMiddleware(api))
-// app.post('/api/forward', asyncMiddleware(forward))
+app.get('/api', asyncMiddleware(api))
+app.post('/api/forward', asyncMiddleware(forward))
 
 app.get('*', (req, res) => {
   res.sendFile(path.resolve(__dirname, '..', 'build', 'index.html'))
