+++ packages/solidarity-client/server/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../../tsconfig.json",
   "compilerOptions": {
-    "outDir": "dist/server"
+    "outDir": "dist"
   }
 }
